move iteration out of the core/__init__.py to core/_neuron.py
* Allows for functions in _neuron to refer to iterators
* taught iter_segments to take a Section()
